requirements.txt
# 코디세이 시험 자동 채점 시스템 의존성

# YAML 파싱 (미션 설정 파일 로드용)
PyYAML>=6.0

# 표준 라이브러리만 사용 (별도 설치 불필요):
# - subprocess: 명령 실행
# - os: 파일 시스템 작업
# - re: 정규표현식
# - json: JSON 처리
# - pathlib: 경로 관리
# - dataclasses: 데이터 클래스
# - typing: 타입 힌트
# - abc: 추상 클래스
